# Optimize command encoding and decoding a bit 