@nussjustin 2019-02-19
Benchmark with `-cpu 8` show smaller improvements but no drastic change for the `no_pipelining` benchmark (which reinforces my guess that it was a hiccup on my test system):

```
name                                  old time/op    new time/op    delta
SerialGetSet/radix-8                    43.9µs ± 1%    44.2µs ± 1%     ~     (p=0.222 n=5+5)
ParallelGetSet/radix/no_pipelining-8    15.4µs ± 6%    15.6µs ± 7%     ~     (p=0.548 n=5+5)
ParallelGetSet/radix/one_pipeline-8     19.0µs ± 5%    17.2µs ± 2%   -9.76%  (p=0.008 n=5+5)
ParallelGetSet/radix/default-8          16.1µs ± 4%    14.1µs ± 4%  -12.17%  (p=0.008 n=5+5)

name                                  old alloc/op   new alloc/op   delta
SerialGetSet/radix-8                     67.0B ± 0%     67.0B ± 0%     ~     (all equal)
ParallelGetSet/radix/no_pipelining-8      115B ± 0%      115B ± 0%     ~     (all equal)
ParallelGetSet/radix/one_pipeline-8      69.4B ± 1%     69.6B ± 1%     ~     (p=1.000 n=5+5)
ParallelGetSet/radix/default-8           70.0B ± 0%     70.0B ± 0%     ~     (all equal)

name                                  old allocs/op  new allocs/op  delta
SerialGetSet/radix-8                      4.00 ± 0%      4.00 ± 0%     ~     (all equal)
ParallelGetSet/radix/no_pipelining-8      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
ParallelGetSet/radix/one_pipeline-8       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
ParallelGetSet/radix/default-8            4.00 ± 0%      4.00 ± 0%     ~     (all equal)
```