@nussjustin 2019-02-19
Rerunning the `no_pipelining` benchmark still gives a small performance decrease, but much less. Still high variation in ns/op for each run.

```
name                                   old time/op    new time/op    delta
ParallelGetSet/radix/no_pipelining-16    16.0µs ± 7%    16.8µs ± 9%   ~     (p=0.310 n=5+5)

name                                   old alloc/op   new alloc/op   delta
ParallelGetSet/radix/no_pipelining-16      116B ± 0%      116B ± 0%   ~     (all equal)

name                                   old allocs/op  new allocs/op  delta
ParallelGetSet/radix/no_pipelining-16      6.00 ± 0%      6.00 ± 0%   ~     (all equal)
```