@mediocregopher 2019-02-19
Hmm that's super interesting! I didn't know that `defer` had any performance impact at all, I thought it was more of a compile-time thing, but I guess it must be putting something on the stack or something. I also wouldn't have expected there to be any change between using MultiWrite vs appending it all to a scratch and using Write, since it's all going into a `bufio.Writer` anyway. But the benchmarks don't lie :P 

Is there any reason to use MultiWrite at all anymore? If using it is inherently worse than using a scratch I'd rather just do that everywhere. If you don't feel like removing it here I can make an issue to remind me to go back and do it when I get home.