@nussjustin 2019-02-20
> I didn't know that `defer` had any performance impact at all, I thought it was more of a compile-time thing, but I guess it must be putting something on the stack or something.

There is a small blog post here that shows the small overhead of defer: https://medium.com/i0exception/runtime-overhead-of-using-defer-in-go-7140d5c40e32

Normally one wouldn't need to think about it, but if you have many of them (and we have), there combined cost becomes easily measurable (just run the default parallel benchmark with -cpuprofile and you will see a few `runtime.defer*` functions). 

> I also wouldn't have expected there to be any change between using MultiWrite vs appending it all to a scratch and using Write, since it's all going into a `bufio.Writer` anyway.

Avoiding interface calls and packing the arguments into a slice (for the variable arguments of `bytes.MultiWrite`) helps tiny bit, but most of it comes from the `*bufio.Writer` not having to check it's available space and decide whether to flush multiple times.

There is a cost though to appending everything to `*scratch`. For small commands like in our benchmarks this cost is less than what we gain but for bigger values (e.g. when setting big strings) appending could be worse than multiple writes and eat more memory and require more allocations (since `*scratch` is rather small).

Maybe we should add a benchmark for this too (`BenchmarkLargeStringArgs`, because naming is hard? ;-)) and see if the overhead is "ok" or if we should keep a special case for large strings or strings larger than what fits into the `*scratch` buffer.

> Is there any reason to use MultiWrite at all anymore? If using it is inherently worse than using a scratch I'd rather just do that everywhere. If you don't feel like removing it here I can make an issue to remind me to go back and do it when I get home.

I wanted to keep this change small and since the other `MarshalRESP` methods are basically never called they won't affect most users. I only updated `BulkStringBytes` as part of this change for symmetry with `BulkString`.