@mediocregopher 2019-02-20
I had some free time so I tested each of the three optimizations individually and they all definitely help, the defer one helps a surprising amount even.

> Maybe we should add a benchmark for this too (`BenchmarkLargeStringArgs`, because naming is hard? ;-)) and see if the overhead is "ok" or if we should keep a special case for large strings or strings larger than what fits into the `*scratch` buffer.

How about `SerialGetSetLargeArgs`, so we can do it for each driver too if we want. But yeah, we might need to do some shenanigans for those which are larger than scratch, it's definitely worth testing out.

> I wanted to keep this change small and since the other `MarshalRESP` methods are basically never called they won't affect most users.

That's fine, I can do the rest at a later time if I get the gumption to do it.