@nussjustin 2019-02-21
Turns out we don't really lose performance for big strings, since `(resp2.BulkString).MarshalRESP` was already appending the string to the `*scratch` slice since it had to convert it to a `[]byte` anyway for `bytesutil.MultiWrite`. Also since we pool our buffers, we often don't need to do any allocations (especially in a non-concurrent program, where we always reuse the same `*scratch` slice).

I added the benchmark and rebased the branch onto master. Here are the results on my notebook:

```
name                                      time/op
SerialGetSetLargeArgs/radix-8             66.8µs ± 0%
SerialGetSetLargeArgs/redigo-8            64.3µs ± 1%
SerialGetSetLargeArgs/redispipe-8          500µs ± 1%
SerialGetSetLargeArgs/redispipe_pause0-8  62.5µs ± 0%

name                                      alloc/op
SerialGetSetLargeArgs/radix-8             12.5kB ± 0%
SerialGetSetLargeArgs/redigo-8            24.7kB ± 0%
SerialGetSetLargeArgs/redispipe-8         13.8kB ± 0%
SerialGetSetLargeArgs/redispipe_pause0-8  13.8kB ± 0%

name                                      allocs/op
SerialGetSetLargeArgs/radix-8               4.00 ± 0%
SerialGetSetLargeArgs/redigo-8              8.00 ± 0%
SerialGetSetLargeArgs/redispipe-8           11.0 ± 0%
SerialGetSetLargeArgs/redispipe_pause0-8    11.0 ± 0%
```

radix isn't too slow, probably because we reuse our buffers and thus avoid allocations after the first benchmark iteration.

We could still fall back to multiple writes when `*scratch` is not big enough to avoid allocations (mostly in cases where `*scratch` is often freshly allocated or just too small), but then we would also need to define some cut off length at which we do this, since allocating a new, bigger slice can still often be faster than multiple writes and amortizes over time as long as the buffer is reused.

We could also just ignore this for now, until it becomes obvious that this is a problem. Since most uses are probably more or less consistent in the size of string arguments, the pooling should work good enough for them.